Feature/product

+++ lib/src/utils/currency_formatted.dart
@@ -1,6 +1,6 @@
 import 'package:intl/intl.dart';
 
-String formatCurrency(double price) {
+String formatCurrency(int price) {
   final format = NumberFormat.decimalPattern('vi_VN');
   return "${format.format(price)}đ"; 
 }
